refactor: ResourceLoader 클래스 유틸화
