+++ tomcat/src/main/java/org/apache/coyote/util/ResourceLoader.java
@@ -1,4 +1,4 @@
-package org.apache.coyote;
+package org.apache.coyote.util;
 
 import java.io.File;
 import java.io.IOException;
@@ -6,9 +6,6 @@ import java.net.URL;
 import java.nio.file.Files;
 
 public class ResourceLoader {
-
-    private static final String ROOT_PATH = "/";
-    private static final String INDEX_PAGE = "static/index.html";
 
     public static byte[] get(final String uri) throws IOException {
         URL url = getUrl(uri);
@@ -22,9 +19,6 @@ public class ResourceLoader {
     }
 
     private static URL getUrl(final String uri) {
-        if (uri.equals(ROOT_PATH)) {
-            return ResourceLoader.class.getClassLoader().getResource(INDEX_PAGE);
-        }
         return ResourceLoader.class.getClassLoader().getResource("static" + uri);
     }
 }
